requirements.txt
playwright>=1.40.0
streamlit>=1.28.0
nest-asyncio>=1.5.8
python-dotenv>=1.0.0
firecrawl-py>=0.0.16
pydantic>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.0

